+++ CHANGELOG.md
@@ -19,6 +19,7 @@ title: Changelog
 * `<cc-env-var-create>`: add a strict mode
 * `<cc-error>`: add a notice mode
 * Update all static assets URLs to Cellar v2
+* `<cc-header-addon>`: display realId in a readonly cc-input-text next to the ID
 
 ### For devs